Modified permission requests
Created a couple methods to handle permission requests and handle if the
user denies the requests such that the camera and gallery will not be
able to be opened otherwise. Before this correction, the app crashed if
permissions were denied and permissions were buggy and would cause the
app to crash and then make requests.